Исправления в командах чтения/записи настроек двигателя

+++ DrillingRig.ConfigApp/AppControl/EngineSettingsSpace/EngineSettingsWritable.cs
@@ -5,7 +5,7 @@ namespace DrillingRig.ConfigApp.AppControl.EngineSettingsSpace {
 		public ushort Inom { get; set; }
 		public ushort Nnom { get; set; }
 		public ushort Nmax { get; set; }
-		public ushort Pnom { get; set; }
+		public uint Pnom { get; set; }
 		public ushort CosFi { get; set; }
 		public ushort Eff { get; set; }
 		public ushort Mass { get; set; }
